+++ modules/sdk-coin-hash/test/unit/hash.ts
@@ -20,6 +20,10 @@ import {
 } from '../resources/hash';
 import should = require('should');
 
+import { loadWebAssembly } from '@bitgo/sdk-opensslbytes';
+
+const openSSLBytes = loadWebAssembly().buffer;
+
 describe('HASH', function () {
   let bitgo: TestBitGoAPI;
   let hash;
@@ -414,13 +418,16 @@ describe('HASH', function () {
     });
 
     it('should recover funds for non-bitgo recoveries', async function () {
-      const res = await thash.recover({
-        userKey: wrwUser.userPrivateKey,
-        backupKey: wrwUser.backupPrivateKey,
-        bitgoKey: wrwUser.bitgoPublicKey,
-        walletPassphrase: wrwUser.walletPassphrase,
-        recoveryDestination: destinationAddress,
-      });
+      const res = await thash.recover(
+        {
+          userKey: wrwUser.userPrivateKey,
+          backupKey: wrwUser.backupPrivateKey,
+          bitgoKey: wrwUser.bitgoPublicKey,
+          walletPassphrase: wrwUser.walletPassphrase,
+          recoveryDestination: destinationAddress,
+        },
+        openSSLBytes
+      );
       res.should.not.be.empty();
       res.should.hasOwnProperty('serializedTx');
       sandBox.assert.calledOnce(thash.getAccountBalance);
@@ -479,35 +486,44 @@ describe('HASH', function () {
 
     it('should throw error if userkey is not present', async function () {
       await thash
-        .recover({
-          backupKey: wrwUser.backupPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          walletPassphrase: wrwUser.walletPassphrase,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            backupKey: wrwUser.backupPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            walletPassphrase: wrwUser.walletPassphrase,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('missing userKey');
     });
 
     it('should throw error if wallet passphrase is not present', async function () {
       await thash
-        .recover({
-          userKey: wrwUser.userPrivateKey,
-          backupKey: wrwUser.backupPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            userKey: wrwUser.userPrivateKey,
+            backupKey: wrwUser.backupPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('missing wallet passphrase');
     });
 
     it('should throw error if there is no balance', async function () {
       await thash
-        .recover({
-          userKey: wrwUser.userPrivateKey,
-          backupKey: wrwUser.backupPrivateKey,
-          bitgoKey: wrwUser.bitgoPublicKey,
-          walletPassphrase: wrwUser.walletPassphrase,
-          recoveryDestination: destinationAddress,
-        })
+        .recover(
+          {
+            userKey: wrwUser.userPrivateKey,
+            backupKey: wrwUser.backupPrivateKey,
+            bitgoKey: wrwUser.bitgoPublicKey,
+            walletPassphrase: wrwUser.walletPassphrase,
+            recoveryDestination: destinationAddress,
+          },
+          openSSLBytes
+        )
         .should.rejectedWith('Did not have enough funds to recover');
     });
   });