+++ modules/sdk-coin-tia/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.16](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-tia@2.1.15...@bitgo/sdk-coin-tia@2.1.16) (2024-08-13)
+
+**Note:** Version bump only for package @bitgo/sdk-coin-tia
+
 ## [2.1.15](https://github.com/BitGo/BitGoJS/compare/@bitgo/sdk-coin-tia@2.1.14...@bitgo/sdk-coin-tia@2.1.15) (2024-08-07)
 
 **Note:** Version bump only for package @bitgo/sdk-coin-tia